+++ internal/storage/entities.go
@@ -16,22 +16,22 @@ var (
 type OrderStatus struct {
 	Number     string    `json:"number"`
 	Status     string    `json:"status"`
-	Amount     int64     `json:"accrual,omitempty"`
+	Amount     float64   `json:"accrual,omitempty"`
 	UploadedAt time.Time `json:"uploaded_at"`
 	UserID     int64     `json:"-"`
 }
 
 type Withdrawn struct {
 	OrderID     string    `json:"order"`
-	Amount      int64     `json:"sum"`
+	Amount      float64   `json:"sum"`
 	ProcessedAt time.Time `json:"processed_at"`
 	UserID      int64     `json:"-"`
 }
 
 type Balance struct {
 	UserID          int64
-	BalanceAmount   int64
-	WithdrawnAmount int64
+	BalanceAmount   float64
+	WithdrawnAmount float64
 }
 
 type User struct {
